refactor(providers): split upload & media providers out of app_providers.dart
First sub-batch of #4506's final wave (9a of 9c). Moves the 10-entity
upload/media cluster to `lib/providers/upload_media_providers.dart`:

Auth/media chain (preserved through batches 5 and 7):
- _BlossomAuthAdapter (private adapter for BlossomAuthProvider)
- _FirebasePerformanceAdapter (private adapter for BlossomPerformanceMonitor)
- blossomAuthService
- mediaViewerAuthServiceProvider (legacy `Provider<X>(...)` style)
- mediaAuthInterceptor

Upload chain:
- blossomUploadService
- uploadManager (top of upload chain — many UI consumers)

API + audio:
- apiService (nip98-backed)
- crosspostApiClient (Bluesky-toggle settings)
- audioPlaybackService

`app_providers.dart` adds `import 'upload_media_providers.dart';`
(back-import — `bugReportService` consumes `blossomUploadServiceProvider`)
and `export 'upload_media_providers.dart';` (keeps 28 external consumer
imports unchanged).

No mutual import with `app_providers.dart` from the new file — all
outbound deps reach already-split files: `auth_providers.dart`
(authService, oauthClient, oauthConfig, nip98AuthService),
`moderation_providers.dart` (ageVerificationService, contentFilterService),
`environment_provider.dart` (currentEnvironmentProvider).

Discovered + fixed during implementation:
- AudioPlaybackService lives in the `sound_service` package, not
  `openvine/services/audio_playback_service.dart` (path that didn't
  exist) — corrected the import.

Removed 9 now-unused imports from `app_providers.dart` via `dart fix`:
blossom_upload_service, api_service, auth_service (no longer needed),
crosspost_api_client, media_auth_interceptor, etc.

Codegen moved: 6 *Provider chunks (~378 LoC) relocated from
`app_providers.g.dart` into `upload_media_providers.g.dart`.

This is the cleanest cluster of the wave — no in-line state machines,
no bridges, no dead providers, no log-level mismatches. Pure
dependency injection.

After this batch, `app_providers.dart` is 1120 → 978 LoC. Two
sub-batches remain (9b: Repositories, 9c: Social + foundation).

Refs #4506, #4339 Wave 2
